+++ test/utils.test.js
@@ -4,7 +4,7 @@ const {
   setDir,
   getLastNumber,
   removeLastNumber,
-  sortFilesByLastNumber,
+  naturalSort,
   autoIncrease,
   resolveOutputFile,
 } = require('../src/utils');
@@ -62,6 +62,14 @@ describe('#getLastNumber', () => {
     expect(getLastNumber('file01.mp4')).toEqual('01');
     expect(getLastNumber('file0_1.mp4')).toEqual('1');
     expect(getLastNumber('file-0150.mp4')).toEqual('0150');
+    expect(getLastNumber('file.mp4')).toEqual('');
+  });
+
+  it('should works', () => {
+    expect(getLastNumber('test (1).txt')).toEqual('1');
+    expect(getLastNumber('test [21].txt')).toEqual('21');
+    expect(getLastNumber('2019_test_{11}.txt')).toEqual('11');
+    expect(getLastNumber('2019 test.txt')).toEqual('');
   });
 });
 
@@ -70,24 +78,74 @@ describe('#removeLastNumber', () => {
     expect(removeLastNumber('dir/filename1.mp4')).toEqual('dir/filename.mp4');
     expect(removeLastNumber('dir/file 01.mp4')).toEqual('dir/file.mp4');
     expect(removeLastNumber('test_00_001.txt')).toEqual('test_00.txt');
+    expect(removeLastNumber('test-01.txt')).toEqual('test.txt');
   });
 
   it('does not changes if path has not last number.', () => {
     expect(removeLastNumber('dir/12_monkeys.mp4')).toEqual('dir/12_monkeys.mp4');
   });
+
+  it('should works also.', () => {
+    expect(removeLastNumber('test (1).txt')).toEqual('test.txt');
+    expect(removeLastNumber('test [21].txt')).toEqual('test.txt');
+    expect(removeLastNumber('2019_test_{11}.txt')).toEqual('2019_test.txt');
+    expect(removeLastNumber('2019 test.txt')).toEqual('2019 test.txt');
+  });
 });
 
-describe('#sortFilesByLastNumber', () => {
+describe('#naturalSort', () => {
   it('works.', () => {
     const files = [
-      'test/video file 01.mp4',
-      'test/video file 04.mp4',
+      'test/video file 1.mp4',
+      'test/video file 14.mp4',
       'test/video file 2.mp4',
-      'test/video file_5.mp4',
-      'test/video file-003.mp4',
+      'test/video file 5.mp4',
+      'test/video file 003.mp4',
     ];
-    const sorted = sortFilesByLastNumber(files);
-    expect(sorted[4]).toEqual('test/video file_5.mp4');
+    const sorted = [
+      'test/video file 1.mp4',
+      'test/video file 2.mp4',
+      'test/video file 003.mp4',
+      'test/video file 5.mp4',
+      'test/video file 14.mp4',
+    ];
+    expect(naturalSort(files)).toEqual(sorted);
+  });
+
+  it('works between different file names.', () => {
+    const files = [
+      'other test 2.mp4',
+      'test 14.mp4',
+      '2 test 11.mp4',
+      'test 2.mp4',
+      'other test 15.mp4',
+      '2 test 3.mp4',
+    ];
+    const sorted = [
+      '2 test 3.mp4',
+      '2 test 11.mp4',
+      'other test 2.mp4',
+      'other test 15.mp4',
+      'test 2.mp4',
+      'test 14.mp4',
+    ];
+    expect(naturalSort(files)).toEqual(sorted);
+  });
+
+  it('works also.', () => {
+    const files = [
+      'test_14.mp4',
+      '2 test (11).mp4',
+      'test_2.mp4',
+      '2 test (3).mp4',
+    ];
+    const sorted = [
+      '2 test (3).mp4',
+      '2 test (11).mp4',
+      'test_2.mp4',
+      'test_14.mp4',
+    ];
+    expect(naturalSort(files)).toEqual(sorted);
   });
 });