+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added new method `naturalSort`.
+
+### Changed
+
+- `getLastNumber` - now available getting last number in brackets. e.g. `test (1).txt`
+- `removeLastNumber` - now available removing last number in brackets. e.g. `test (1).txt`
+
+### Deprecated
+
+- `sortFilesByLastNumber` was deprecated. use `naturalSort` instead.
+
 ## [1.0.1] - 2019-12-01
 
 - Fixed `setDir` (31e24810424f5a6de6159848fa0b1e22a66930c7)